+++ examples/receive_tcp_stream.py
@@ -1,8 +1,11 @@
+import time
+
 import gbvision as gbv
 
 
 def main():
-    receiver = gbv.TCPStreamReceiver('127.0.0.1', 5808)
+    receiver = gbv.AsyncTCPStreamReceiver('127.0.0.1', 5808)
+    time.sleep(5)
     window = gbv.StreamWindow(window_name='stream example', wrap_object=receiver)
     window.show()
