+++ examples/show_edges.py
@@ -4,7 +4,7 @@ import time
 
 def main():
     camera = gbv.AsyncUSBCamera(0)
-    time.sleep(5)  # wait for read thread to start
+    camera.wait_start_reading()
     camera.set_frame_size(640, 480)
     orig_window = gbv.CameraWindow('original', camera)
     edges_window = gbv.CameraWindow('edges', camera, drawing_pipeline=gbv.edges + gbv.gray)
