+++ gbvision/__init__.py
@@ -91,6 +91,7 @@ from .utils.pipeline import PipeLine
 from .utils.game_object import GameObject
 from .utils.tracker import Tracker
 from .utils.readable import Readable
+from .utils.async_readable import AsyncReadable
 
 
 # configure opencv
